+++ packages/entities/entities-plugins/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [9.28.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.27.1...@kong-ui-public/entities-plugins@9.28.0) (2025-07-04)
+
+
+### Features
+
+* **entities-*:** migrate plugin icon to a standalone repo [KM-1318] ([#2158](https://github.com/Kong/public-ui-components/issues/2158)) ([6157ecb](https://github.com/Kong/public-ui-components/commit/6157ecb1bb7e779193f43fe1265ba466d601e034))
+
+
+
+
+
 ## [9.27.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.27.0...@kong-ui-public/entities-plugins@9.27.1) (2025-06-26)
 
 **Note:** Version bump only for package @kong-ui-public/entities-plugins